build search feature

+++ src/main/java/com/example/SearchEngine/invertedIndex/TrieInvertedIndex.java
@@ -1,8 +1,9 @@
 package com.example.SearchEngine.invertedIndex;
 
 import com.example.SearchEngine.Analyzers.Analyzer;
-import com.example.SearchEngine.Analyzers.AnalyzerEnum;
 import com.example.SearchEngine.Tokenization.Token;
+import com.example.SearchEngine.invertedIndex.utility.CollectionInfo;
+import com.example.SearchEngine.invertedIndex.utility.SchemaAnalyzer;
 import com.example.SearchEngine.schema.service.SchemaDefaultService;
 import com.example.SearchEngine.schema.util.SchemaRoot;
 import com.example.SearchEngine.utils.storage.service.SchemaPathService;
@@ -11,7 +12,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 @Service
 public class TrieInvertedIndex implements InvertedIndex {
@@ -25,11 +28,20 @@ public class TrieInvertedIndex implements InvertedIndex {
     @Autowired
     private ObjectMapper mapper;
 
-    private Analyzer getFieldAnalyzer(Map<String, Object> schemaField) {
-        return AnalyzerEnum.EnglishAnalyzer.getAnalyzer();
+
+    public boolean checkWordExist(TrieNode root, Token token) {
+        TrieNode currentTrieNode = root;
+        for (int i = 0; i < token.getWord().length(); i++) {
+            Character c = token.getWord().charAt(i);
+            if (!currentTrieNode.hasNextNode(c)) {
+                return false;
+            }
+            currentTrieNode = currentTrieNode.getNextNode(c);
+        }
+        return true;
     }
 
-    private TrieNode getWordLastNode(TrieNode root, Token token) {
+    public TrieNode getWordLastNode(TrieNode root, Token token) {
         TrieNode currentTrieNode = root;
         for (int i = 0; i < token.getWord().length(); i++) {
             Character c = token.getWord().charAt(i);
@@ -37,6 +49,7 @@ public class TrieInvertedIndex implements InvertedIndex {
         }
         return currentTrieNode;
     }
+
 
     private void indexer(TrieNode root, Integer documentId, String fieldName, List<Token> tokens) {
         for (Token token : tokens) {
@@ -48,18 +61,22 @@ public class TrieInvertedIndex implements InvertedIndex {
 
     public void remover(List<Token> tokens, TrieNode root, Integer documentId) {
         for (Token token : tokens) {
-            TrieNode lastNode = getWordLastNode(root, token);
-            lastNode.deleteDocument(documentId);
+            if (checkWordExist(root, token)) {
+                TrieNode lastNode = getWordLastNode(root, token);
+                lastNode.deleteDocument(documentId);
+            }
         }
     }
 
     @Override
     public void addDocument(String schemaName, Map<String, Object> document) throws Exception {
         TrieNode root = schemaRoot.getSchemaRoot(schemaName);
-        Map<String, Object> schema = (Map<String, Object>) schemaDefaultService.getSchema(schemaName).get("properties");
-
+        CollectionInfo.updateNumberOfDocument(schemaName, 1);
         for (String fieldName : document.keySet()) {
-            List<Token> tokens = getTokens(fieldName, schema, document);
+            List<Token> tokens = getTokens(schemaName, fieldName, document);
+            if (!tokens.isEmpty()) {
+                CollectionInfo.addField(schemaName, (Integer) document.get("id"), tokens);
+            }
             indexer(root, (Integer) document.get("id"), fieldName, tokens);
         }
     }
@@ -67,23 +84,27 @@ public class TrieInvertedIndex implements InvertedIndex {
     @Override
     public void deleteDocument(String schemaName, Integer documentId) throws Exception {
         String path = schemaPathService.getSchemaPath(schemaName) + "documents/" + documentId;
+        CollectionInfo.updateNumberOfDocument(schemaName, -1);
         Map<String, Object> document = mapper.readValue(new File(path), Map.class);
         TrieNode root = schemaRoot.getSchemaRoot(schemaName);
-        Map<String, Object> schema = (Map<String, Object>) schemaDefaultService.getSchema(schemaName).get("properties");
 
         for (String fieldName : document.keySet()) {
-            List<Token> tokens = getTokens(fieldName, schema, document);
+            List<Token> tokens = getTokens(schemaName, fieldName, document);
+            if (!tokens.isEmpty()) {
+                CollectionInfo.removeField(schemaName, (Integer) document.get("id"), tokens);
+            }
             remover(tokens, root, documentId);
         }
     }
 
-    private List<Token> getTokens(String fieldName, Map<String, Object> schema, Map<String, Object> document) {
-        Analyzer analyzer;
+    private List<Token> getTokens(String schemaName, String fieldName, Map<String, Object> document) throws Exception {
         List<Token> tokens = new ArrayList<>();
+        Analyzer analyzer = SchemaAnalyzer.getAnalyzer(schemaName);
+        Map<String, Object> schema = (Map<String, Object>) schemaDefaultService.getSchema(schemaName).get("properties");
+
         if (!fieldName.equals("id")) {
             Map<String, Object> schemaField = (Map<String, Object>) schema.get(fieldName);
             if (schemaField.get("type").equals("text")) {
-                analyzer = getFieldAnalyzer(schemaField);
                 Double weight = 1.0;
                 if (schemaField.containsKey("weight")) {
                     weight = (Double) schemaField.get("weight");
@@ -93,4 +114,5 @@ public class TrieInvertedIndex implements InvertedIndex {
         }
         return tokens;
     }
+
 }
